Notify vcmi-client that the server has stopped.
When winning the campaign there is a lock that waits for the old server to stop before allowing to continue with restarting the server. While the lock was released on PC builds as the server runs differently on Android this was missed making campaigns always hang the app when you win.

This commit makes sure we notify the client if the server stops so it can release the lock.

This commit also fixes the issue that the server's connection was never unbound, hence calling bind again did nothing. The main issue was that since the callback was created on the stack, but was only stored as a WeakReference it immediately became garbage collected, so the link to call unbind was gone.