Add support for extra resolutions built in
The current package only supports 800x600, and adding support for other resolutions is non-trivial on Android. This change makes sure that the essential package that is built contains the latest VCMI essential mod 1.01 from https://wiki.vcmi.eu/Mod_list#Utilities

The changes in the update_internal_assets.py are required to make sure each file is only addedd to the zipfile once - otherwise the python package allows the same file to be added to the zip multiple times causing issues during decompression.